Better way of parsing module descriptions

+++ Rakefile
@@ -1,26 +1,48 @@
-require 'yaml'
+require 'YAML'
+require 'JSON'
+
+class Formula
+  def method_missing(method_name, *args); end
+  def self.method_missing(method_name, *args); end
+  def build; Formula.new; end
+  def self.build; Formula.new; end
+  def self.test; end
+end
 
 namespace :modules do
 
-    desc "Build _data directory"
-    task :generate do
-      tmp_dir = "/tmp/homebrew-nginx-generate"
-      git_url = "https://github.com/marcqualie/homebrew-nginx.git"
-      if Dir.exists? tmp_dir
-        Dir.chdir(tmp_dir) do
-          sh "git fetch origin"
-          sh "git pull origin master"
-        end
-      else
-        sh "git clone #{git_url} #{tmp_dir}"
+  desc "Build _data directory"
+  task :generate do
+    tmp_dir = "/tmp/homebrew-nginx-generate"
+    git_url = "https://github.com/marcqualie/homebrew-nginx.git"
+    if Dir.exists? tmp_dir
+      Dir.chdir(tmp_dir) do
+        sh "git fetch origin"
+        sh "git pull origin master"
       end
-      modules = []
-      Dir["#{tmp_dir}/*-module.rb"].each do |filename|
+    else
+      sh "git clone #{git_url} #{tmp_dir}"
+    end
+
+    module_data = {}
+    modules = []
+
+    $LOAD_PATH.unshift File.expand_path(tmp_dir)
+    FileUtils.touch "#{tmp_dir}/formula.rb"
+    require "#{tmp_dir}/nginx-full"
+    NginxFull.core_modules.each do |arr|
+      module_data[arr[0]] = arr[2]
+    end
+    NginxFull.third_party_modules.each do |name, desc|
+      module_data[name] = desc
+    end
+
+    module_data.each do |name, desc|
+      filename = "#{tmp_dir}/#{name}-nginx-module.rb"
+      if File.exists? filename
         contents = File.open(filename) { |f| f.read }
-        match = /\/([a-z0-9\-]+)\-nginx\-module\.rb/.match(filename)
-        name = match[1]
-        match = /url [\"'](.*)[\"']/.match(contents)
-        url = match[1]
+        match = /homepage [\"'](.*)[\"']/.match(contents)
+        homepage = match[1]
         match = /sha1 [\"'](.*)[\"']/.match(contents)
         sha1 = match[1]
         match = /version [\"'](.*)[\"']/.match(contents)
@@ -29,17 +51,26 @@ namespace :modules do
           match = /\/v?([0-9\.]+)\.tar\.gz/.match(contents)
           version = match[1] if match
         end
-        version = sha1[0..6] unless version
-        modules.push({
-          'name' => name,
-          'version' => version,
-          'sha1' => sha1,
-          'description' => ""
-        })
+      else
+        sha1 = nil
+        homepage = nil
+        version = 'master'
       end
-      File.open('_data/modules.yml', 'w') do |f|
-        YAML.dump(modules, f)
-      end
+      version = sha1[0..6] unless version
+      modules.push({
+        'name' => name,
+        'homepage' => homepage,
+        'version' => version,
+        'sha1' => sha1,
+        'description' => desc
+      })
     end
+    modules.sort_by! do |m|
+      m['name']
+    end
+    File.open('_data/modules.yml', 'w') do |f|
+      YAML.dump(modules, f)
+    end
+  end
 
 end
